refactor(medications): apply conventions
