--- src/app/medications/components/MedicationIconPicker.module.css
@@ -1,20 +1,0 @@
-.icon_group {
-  max-width: 565px;
-}
-
-.icon_button {
-  background-color: transparent;
-  border: none;
-
-  margin: 5px;
-  padding: 5px;
-}
-
-.icon_button:hover {
-  background-color: #83b3f3;
-}
-
-.icon_button_selected {
-  background-color: #83b3f3;
-  border: none;
-}
